scripts/requirements.txt
requests>=2.31.0
PyYAML>=6.0
beautifulsoup4>=4.12.0
packaging>=23.0
